+++ include/class.search.php
@@ -41,7 +41,7 @@ abstract class SearchBackend {
     );
 
     abstract function update($model, $id, $content, $new=false, $attrs=array());
-    abstract function find($query, QuerySet $criteria);
+    abstract function find($query, QuerySet $criteria, $addRelevance=true);
 
     static function register($backend=false) {
         $backend = $backend ?: get_called_class();
@@ -76,9 +76,9 @@ class SearchInterface {
         $this->bootstrap();
     }
 
-    function find($query, QuerySet $criteria) {
+    function find($query, QuerySet $criteria, $addRelevance=true) {
         $query = Format::searchable($query);
-        return $this->backend->find($query, $criteria);
+        return $this->backend->find($query, $criteria, $addRelevance);
     }
 
     function update($model, $id, $content, $new=false, $attrs=array()) {
@@ -324,7 +324,7 @@ class MysqlSearchBackend extends SearchBackend {
         return implode(' ', $results);
     }
 
-    function find($query, QuerySet $criteria) {
+    function find($query, QuerySet $criteria, $addRelevance=true) {
         global $thisstaff;
 
         $criteria = clone $criteria;
@@ -333,26 +333,57 @@ class MysqlSearchBackend extends SearchBackend {
         // If using boolean operators, search in boolean mode. This regex
         // will ensure proper placement of operators, whitespace, and quotes
         // in an effort to avoid crashing the query at MySQL
-        if (preg_match('/^(?:[(+<>~-]*(\w+[*]?|"[^"]+")[)]?(\s+|$))+$/u', $query, $T = array()))
+        $query = $this->quote($query);
+        if (preg_match('/(?=(?:^|\s)[()"+<>~-])\s*(?:[(+<>~-]*(\w+[*]?|"[^"]+")[)]?(\s+|$))+$|\w[*]/u', $query, $T = array()))
             $mode = ' IN BOOLEAN MODE';
         #elseif (count(explode(' ', $query)) == 1)
         #    $mode = ' WITH QUERY EXPANSION';
-        $query = $this->quote($query);
         $search = 'MATCH (Z1.title, Z1.content) AGAINST ('.db_input($query).$mode.')';
 
         switch ($criteria->model) {
         case false:
         case 'TicketModel':
+            if ($addRelevance) {
+                $criteria = $criteria->extra(array(
+                    'select' => array(
+                        '__relevance__' => 'Z1.`relevance`',
+                    ),
+                ));
+            }
+            $criteria->extra(array(
+                'tables' => array(
+                    str_replace(array(':', '{}'), array(TABLE_PREFIX, $search),
+                    "(SELECT COALESCE(Z3.`object_id`, Z5.`ticket_id`, Z8.`ticket_id`) as `ticket_id`, {} AS `relevance` FROM `:_search` Z1 LEFT JOIN `:thread_entry` Z2 ON (Z1.`object_type` = 'H' AND Z1.`object_id` = Z2.`id`) LEFT JOIN `:thread` Z3 ON (Z2.`thread_id` = Z3.`id` AND Z3.`object_type` = 'T') LEFT JOIN `:ticket` Z5 ON (Z1.`object_type` = 'T' AND Z1.`object_id` = Z5.`ticket_id`) LEFT JOIN `:user` Z6 ON (Z6.`id` = Z1.`object_id` and Z1.`object_type` = 'U') LEFT JOIN `:organization` Z7 ON (Z7.`id` = Z1.`object_id` AND Z7.`id` = Z6.`org_id` AND Z1.`object_type` = 'O') LEFT JOIN :ticket Z8 ON (Z8.`user_id` = Z6.`id`) WHERE {}) Z1"),
+                )
+            ));
+            $criteria->filter(array('ticket_id'=>new SqlCode('Z1.`ticket_id`')));
+            break;
+
+        case 'User':
             $criteria->extra(array(
                 'select' => array(
                     '__relevance__' => 'Z1.`relevance`',
                 ),
                 'tables' => array(
                     str_replace(array(':', '{}'), array(TABLE_PREFIX, $search),
-                        "(SELECT COALESCE(Z3.`object_id`, Z5.`ticket_id`) as `ticket_id`, {} AS `relevance` FROM `:_search` Z1 LEFT JOIN `:thread_entry` Z2 ON (Z1.`object_type` = 'H' AND Z1.`object_id` = Z2.`id`) LEFT JOIN `:thread` Z3 ON (Z2.`thread_id` = Z3.`id` AND Z3.`object_type` = 'T') LEFT JOIN `:ticket` Z5 ON (Z1.`object_type` = 'T' AND Z1.`object_id` = Z5.`ticket_id`) WHERE {}) Z1"),
+                    "(SELECT Z6.`id` as `user_id`, {} AS `relevance` FROM `:_search` Z1 LEFT JOIN `:user` Z6 ON (Z6.`id` = Z1.`object_id` and Z1.`object_type` = 'U') LEFT JOIN `:organization` Z7 ON (Z7.`id` = Z1.`object_id` AND Z7.`id` = Z6.`org_id` AND Z1.`object_type` = 'O') WHERE {}) Z1"),
                 )
             ));
-            $criteria->filter(array('ticket_id'=>new SqlCode('Z1.`ticket_id`')))->distinct('ticket_id');
+            $criteria->filter(array('id'=>new SqlCode('Z1.`user_id`')));
+            break;
+
+        case 'Organization':
+            $criteria->extra(array(
+                'select' => array(
+                    '__relevance__' => 'Z1.`relevance`',
+                ),
+                'tables' => array(
+                    str_replace(array(':', '{}'), array(TABLE_PREFIX, $search),
+                    "(SELECT Z2.`id` as `org_id`, {} AS `relevance` FROM `:_search` Z1 LEFT JOIN `:organization` Z2 ON (Z2.`id` = Z1.`object_id` AND Z1.`object_type` = 'O') WHERE {}) Z1"),
+                )
+            ));
+            $criteria->filter(array('id'=>new SqlCode('Z1.`org_id`')));
+            break;
         }
 
         // TODO: Ensure search table exists;
@@ -737,6 +768,18 @@ class SavedSearch extends VerySimpleModel {
 #            ':org' =>        new OrganizationChoiceField(array(
 #                'label' => __('Organization'),
 #            )),
+            ':closed' =>     new DatetimeField(array(
+                'id' => 3204,
+                'label' => __('Closed Date'),
+            )),
+            ':thread__lastresponse' => new DatetimeField(array(
+                'id' => 3205,
+                'label' => __('Last Response'),
+            )),
+            ':thread__lastmessage' => new DatetimeField(array(
+                'id' => 3206,
+                'label' => __('Last Message'),
+            )),
             ':source' =>     new TicketSourceChoiceField(array(
                 'id' => 3201,
                 'label' => __('Source'),
@@ -870,8 +913,10 @@ class SavedSearch extends VerySimpleModel {
                         $filter->add(array('entries__answers__field_id' => $id));
                         break;
                     }
-                    $OP = $other_paths[$type];
-                    $name = $OP . $column;
+                    if ($OP = $other_paths[$type])
+                        $name = $OP . $column;
+                    else
+                        $name = substr($name, 1);
                 }
             }
 
@@ -1107,11 +1152,11 @@ class TicketFlagChoiceField extends ChoiceField {
 class TicketSourceChoiceField extends ChoiceField {
     function getChoices() {
         return array(
-            'w' => __('Web'),
-            'e' => __('Email'),
-            'p' => __('Phone'),
-            'a' => __('API'),
-            'o' => __('Other'),
+            'web' => __('Web'),
+            'email' => __('Email'),
+            'phone' => __('Phone'),
+            'api' => __('API'),
+            'other' => __('Other'),
         );
     }