+++ include/staff/templates/thread-entries.tmpl.php
@@ -4,6 +4,13 @@ $events = $events->getIterator();
 $events->rewind();
 $event = $events->current();
 $htmlId = $options['html-id'] ?: ('thread-'.$this->getId());
+
+$thread_attachments = array();
+foreach (Attachment::objects()->filter(array(
+    'thread_entry__thread__id' => $this->getId(),
+))->select_related('thread_entry', 'file') as $att) {
+    $thread_attachments[$att->object_id][] = $att;
+}
 ?>
 <div id="<?php echo $htmlId; ?>">
     <div id="thread-items" data-thread-id="<?php echo $this->getId(); ?>">
@@ -58,16 +65,16 @@ $htmlId = $options['html-id'] ?: ('thread-'.$this->getId());
         // Set inline image urls.
         <?php
         $urls = array();
-        foreach (AttachmentFile::objects()->filter(array(
-            'attachments__thread_entry__thread__id' => $this->getId(),
-            'attachments__inline' => true,
-        )) as $file) {
-            $urls[strtolower($file->getKey())] = array(
-                'download_url' => $file->getDownloadUrl(),
-                'filename' => $file->name,
+        foreach ($thread_attachments as $eid=>$atts) {
+            foreach ($atts as $A) {
+                if (!$A->inline)
+                    continue;
+                $urls[strtolower($A->file->getKey())] = array(
+                    'download_url' => $A->file->getDownloadUrl(),
+                    'filename' => $A->getFilename(),
                 );
-
             }
+        }
         ?>
         $('#'+container).data('imageUrls', <?php echo JsonDataEncoder::encode($urls); ?>);
         // Trigger thread processing.